requirements_docs.txt
-r requirements_dev.txt
django>=1.11
celery>=4.2.0
django-chamber>=0.5.0
